Get ANGLE static translation libraries to build with mingw.
This change includes all the changes necessary to build the ANGLE
shader translation static libraries with mingw for the emulator,
including:
- Minor patches and definition files.
- The removal of the <algorithm> include in a number of locations
  where it is confusing the compiler after being previously defined
  in C code.

In addition, this change disables GL, D3D, and HLSL from building,
as these require mingw 5.3.0. Disabling these does not prevent the
desired translation libraries from building, but does disable the
overall ANGLE build, which will be desired in the future.

Change-Id: I7605b455466e14d602753f1582c9c770351be0dc