[GLESv3] Don't layout(std140) -> layout(binding = 0)
Only exposed on OpenGL core profiles / older OpenGL driver versions
than 4.3, as "binding" is not a thing pre-4.3
and there is a shader compile error.

This CL just initiailizes fields of the struct with '-1'
"N/A" values instead of relying on create(), so then
the GLSL backend doesn't think "binding" needs
to be generated.

Change-Id: Ib3a08a932c2e0e4ad9156681e44e5e3c8c407992