Emulator boot animation in angle_indirect
There is use cases in emulator where the read framebuffer format is not
compatible with destination texture formats in glCopyTexSubImage.

We remove some of the validation code. We are already doing similar
thing in GLES2. Please revert this CL and the GLES2 CL when we have a
proper fix in the emulator code.

Link to GLES2 part.

https://android.googlesource.com/platform/external/angle/+/emu-master-dev/src/libANGLE/validationES2.cpp#575

BUG: 64540066
Change-Id: I1210480e529d50d0e49a405d0c886671d0dd08ea