Add angle_end2end_tests_app helper APK
Add the ANGLE end2end test APK angle_end2end_tests_app.

This APK is essentially a thin wrapper about the end2end test library,
passing the necessary arguments to test the system EGL. We can't build
and use the full gtest APK in Android, because it depends on too much
external config/infra handled by the gtest Python scripts.

To begin with, this is hard-coded to only run a single test (via
--gtest_filter) to allow us to get all the pieces in place before
running the full suite with filtered by the expectations file. This also
improves iteration time, since the single test runs in a few seconds
compared to tens of minutes for the full suite.

To manually build, install, run, and inspect the output of the helper
APK:

1. Build the APK
  $ m angle_end2end_tests_app
2. Install the APK
  $ adb install -r -d -g --force-queryable \
  testcases/angle_end2end_tests_app/x86_64/angle_end2end_tests_app.apk
3. Run as root
  $ adb root
4. Log user 10 out/in
  $ adb shell am logout-user 10 && sleep 3 && adb shell am switch-user 10
5. Push angle_end2end_tests_expectations.txt
  $ adb push android/cts/angle_end2end_tests_expectations.txt
  /tmp/angle_end2end_tests_expectations.txt
  $ adb shell cp /tmp/angle_end2end_tests_expectations.txt
  /data/user/10/com.android.angle.test/chromium_tests_root/src/tests/angle_end2end_tests_expectations.txt
6. Launch the APK
  $ adb shell am start com.android.angle.test/.AngleNativeTest
7. View the logs and results
  $ adb shell cat /data/media/10/Download/out.txt
8. View the JSON results
  $ adb shell cat /data/media/10/Download/output.json

Bug: b/279980674
Test: Manual (see above)
Change-Id: I2548f86c0fafb69705d51d4a39f67dd41f059dc9