Write log files to /sdcard
Move the AngleNativeTest logging from /sdcard/Download to /sdcard. When
accessing the files via adb, this maps to: /data/media/[0|10].

This makes life easier when kicking off the tests and accessing the log
and results files with hostside runner.

This CL also cleans up any old out.txt files, so the logs are always
fresh and relevant to the current test invocation.

Bug: b/279980674
Test: m AngleEnd2EndTestsHelper
Change-Id: Ied8ba51ea90aef193c04986bb335afdeab8a062d