cts: Convert AngleEnd2EndTestsHelper to instrumented test
Convert AngleEnd2EndTestsHelper to an instrumented test. This allows us
to greatly simplify things on the device and on the host, by making the
test execution much more synchronous.

Specifically, this allows us to remove polling for the output.json file
on the host as well as detecting test (app) crashes immediately, rather
than timing out after not seeing the output.json file in time.

The class AngleEnd2EndTest is the instrumentation entry point (via
AndroidJUnitRunner). This class creates an instance of AngleNativeTest
and launches it, so it is fully initialized from a normal
app-perspective. Next, it calls runEnd2EndTests() directly and blocks
until test completion, either good or bad. If the tests
(AngleNativeTest) crash or exit unexpectedly for any reason,
AngleEnd2EndTest will also return the bad status to the hostside runner.

Bug: b/279980674
Test: am instrument -w
com.android.angle.test/androidx.test.runner.AndroidJUnitRunner

Change-Id: Iad5589b39588eb5d5d879695eb92defe25c17e09