cts: Collect all available log files separately
Currently, collectDeviceLogs() attempts to read all the log files in a
single try-catch block. However, if any file is missing, an exception is
generated and no log files are collected. This means that if there is an
error and the tests are not run at all (e.g., bad expectations file
format), collectDeviceLogs() will fail because output.json is missing,
and we don't get the out.txt log file to debug with.

Separate out the log file data collection so we collect as many of the
log files as are available. Don't return when an error is detected; just
log it and continue on to attempt to collect any other files.

This CL also contains some linter-type cleanup related to Strings.

Bug: b/279980674
Test: atest CtsAngleEnd2EndTests
Change-Id: I4f781acd974db88a7f3bf490773a9c2f941e5e39