cts: Always collect device logs
Always collect and attempt to parse the available logs by moving
runDeviceTests() and collectDeviceLogs() inside a try-catch-finally
block.

If an end2end test crashes, runDeviceTests() will generate a
AssertionError exception, which causes AngleEnd2EndHostTest::run() to
return immediately if runDeviceTests() is not inside a try block. This
means we don't currently collect any logs, so we can't determine which
test crashed to debug it further.

Move runDeviceTests() inside a try-catch-finally block. The only
exception that will be caught is DeviceNotAvailableException, because
that indicates we cannot collect any logs anyway. Inside the finally
block, collect the test logs and attempt to parse the results if the
JSON file is available.

Bug: b/279980674
Test: atest CtsAngleEnd2EndTests
Change-Id: I722595628eaed489d98e6514df5ee00687ca2bf1