cts: Implement ITestCollector
The test suite v2/wear-engprod/cts_device_collect_test is failing when
running CtsAngleEnd2EndTests:

    class com.android.angle.cts.AngleEnd2EndHostTest does not implement
    ITestCollector

The interface ITestCollector only declares a single function:

    /**
     * Enables or disables test collection mode
     * @param shouldCollectTest
     */
    public void setCollectTestsOnly(boolean shouldCollectTest);

Update AngleEnd2EndHostTest to implement ITestCollector with an empty
setCollectTestsOnly(), which returns the list of tests. However, this
also *runs* the tests, which we don't want. For EOS this will return an
empty list which is accurate, because EOS doesn't use ANGLE as the
OpenGL ES driver, so the CtsAngleEnd2EndTests is skipped anyway.

b/432021211 is tracking how to toggle the '--list-tests' argument in the
helper APK and parsing the output.

Bug: b/438878222
Bug: b/432021211
Test: atest CtsAngleEnd2EndTests --collect-tests-only
Change-Id: Ib6e3a92478492b579770bc4c0649f2b1211cfef6