Convert atest cmdline filter to gtest filter
Implement ITestFilterReceiver obj to convert atest cmdline filter
string to combined gtest filter string being passed along to the
target side AngleNativeTest class.

The conversion always appends `*` (wildcard) character to the end of
each filter that specifies the prefix string of a class or a method
for testing, hence all test cases start with the common prefix string
are included.

Bug: b/431773954
Test: Pixel 9 Pro Fold (Comet) - w/o checking ANGLE as default GL lib
Test: atest CtsAngleEnd2EndTestCases - completed w/ expected CRASH
Test: atest CtsAngleEnd2EndTestCases:ClearTestES3.DrawAfterClearWith - 4
Test: atest CtsAngleEnd2EndTestCases:ClearTestES3.DrawAfterClearWith ctsAngleEnd2EndTestCases:Hui - 4
Test: atest CtsAngleEnd2EndTestCases:ClearTestES3 - found 212 tests
Test: atest CtsAngleEnd2EndTestCases:ClearTestES3. - atest failed
Test: atest CtsAngleEnd2EndTestCases:ClearTestES3* - no test found
Change-Id: I89a466adb9e3cc0e2dd641db498d148a0474f4b9