Select ANGLE if vendor API level suffices
When ANGLE is not set as system GL driver, check vendor API level to
decide whether or not to select ANGLE as the GL driver for the test,
and then restore the ANGLE selection settings after the test is done.

The minimum vendor API level is 202604 (ro.vendor.api_level) to target
Android 17 and onward.

The settings of ANGLE selection (package name and driver name) can be
absent and system property prints "null" string as the output. To
restore, the CL keeps the row but replaces its value with "" (empty
string). For other cases, the original values are restored, or the
setting rows are left untouched if ANGLE is the system GL driver.

Remove TODO comment text of b/431804941.

Bug: b/431804941
Test: Android 16 Comet, ro.vendor.api_level=34
Test: ro.board.api_level=202504, ro.build.version.sdk=36
Test: (no test results) atest CtsAngleEnd2EndTestCases
Test: w/ empty ANGLE sel pkgs/vals, restored correctly with "".
Test: w/ xyz,abc/default ANGLE sel pkgs/vals, restored correctly
Test: if ANGLE is sys GL, does nothing on settings restore.
Change-Id: Ib03d5dedce2d29fcbe107f416302429d76a860fe