Implement --collect-tests-only flag
Convert atest --collect-tests-only flag to --list-tests flag of ANGLE
GoogleTest E2E launcher.

As the ANGLE E2E dry-runs w/ --list-tests flag and outputs the list of
test cases to stdout. No output.json will be available in this case.
Hence, the CL parses stdout log and produces a output.json to report to
TradeFed the number of test cases.

Bug: b/432021211
Test: atest CtsAngleEnd2EndTestCases --collect-tests-only
Change-Id: I5d6ffabe62bc8df6da28510bea658b1fba4e5363