Clean up settings restore function
When a setting does not exist, a literal string "null" is printed and
returned. Which may confuse with the circumstance that the setting was
actually set to the value of literal string "null" (not a real world
valid case though).

Changes to explicitly treat the returned value "null" as the setting
was not set and return "\"\"" (an empty string, a workaround due to
the lack of delete API) for the restore function to work w/o handling
the "null" special case.

As well as ignore any restore if the value is a null object, such as
that in the circumstance that ANGLE is set to be the system GL
driver, hence settings were left untouched.

Bug: b/431804941
Test: atest CtsAngleEnd2EndTestCases
Test: w/ empty ANGLE sel pkgs/vals, restored correctly with "".
Test: w/ xyz,abc/default ANGLE sel pkgs/vals, restored correctly
Test: if ANGLE is sys GL, does nothing on settings restore.
Change-Id: I153f9ce5ae76e15b78a08bd8d3e17ff76d267c4c