Move ANGLE test names into multiple classes of pkg
Instead of using package and class names of the CTS Java class as the
only single class name (in ATI POV), split the reported ANGLE E2E test
names at the first `.` character. The first part becomes the class
name, and rest of string becomes the method name in ATI.

The CL helps improve the page loading efficiency of the test results
shown on Android Test Investigate, if the number of test names is
large (e.g. >= 13000).

Bug: b/455869056
Test: atest CtsAngleEnd2EndTestCases
Change-Id: Id92959a3e523abf0e2ffb852293ce522f4ba53cf